fix bug?

+++ ci.sh
@@ -1,15 +1,28 @@
 #!/usr/bin/env bash
-wget -q https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O Miniconda_latest.sh
-bash Miniconda_latest.sh -b -p $HOME/miniconda
-export PATH="$HOME/miniconda/bin:$PATH"
-rm Miniconda_latest.sh
-conda config --set always_yes yes --set show_channel_urls yes
-conda update conda
-conda config --add channels BjornFJohansson
-conda create -qy -n testenv python=3.6 nbval pytest lxml requests beautifulsoup4
-source activate testenv
+if [[ $CI = true ]]||[[ $CI = True ]]
+then
+    echo "Running on CI server"
+    wget -q https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O Miniconda_latest.sh
+    bash Miniconda_latest.sh -b -p $HOME/miniconda
+    export PATH="$HOME/miniconda/bin:$PATH"
+    rm Miniconda_latest.sh
+    conda config --set always_yes yes --set show_channel_urls yes
+    conda update conda
+    conda config --add channels BjornFJohansson
+else
+    echo "Not running on CI server, probably running on local computer"
+    local_computer=true
+fi
+
+conda create -y -n dDEL_testenv python=3.6 nbval pytest lxml requests beautifulsoup4
+source activate dDEL_testenv
 which python
 python --version
 conda install pydna
 conda install pygenome
 python run_test.py
+if [[ $local_computer = true ]]
+then
+    source activate bjorn36
+    conda remove -n dDEL_testenv --all -q
+fi
